refactor(dashboard): extrai seleção de subprojetos e ambiente para módulos compartilhados
Elimina a duplicação de código entre os cards Lançamentos e Branches e
Builds, apontada pela análise estática.

- Adiciona hook useSubprojectSelection com a lógica de seleção de subprojeto
- Adiciona componentes SubprojectSelect e EnvironmentSelect reutilizáveis
- Faz JenkinsJob, JenkinsJobCard e Releases consumirem os módulos compartilhados
