+++ src/components/dashboard/JenkinsJob/index.tsx
@@ -1,23 +1,15 @@
 "use client";
 
-import { useEffect, useMemo, useState } from "react";
 import JenkinsBranchBuildsCard from "@/components/dashboard/JenkinsBranchBuildsCard";
+import { SubprojectSelect } from "@/components/dashboard/SubprojectSelect";
 import { useJenkinsMetrics } from "@/hooks/useJenkinsMetrics";
+import { useSubprojectSelection } from "@/hooks/useSubprojectSelection";
 import { cn } from "@/lib/utils";
-import {
-    Select,
-    SelectContent,
-    SelectItem,
-    SelectTrigger,
-    SelectValue,
-} from "@/components/ui/select";
 import type { JenkinsSubproject } from "@/types/jenkinsSubproject";
 import {
     getJenkinsEnvironmentForDeploy,
     type DeployEnvironment,
 } from "@/types/deployEnvironment";
-
-const EMPTY_SUBPROJECTS: JenkinsSubproject[] = [];
 
 type Props = {
     readonly project: string;
@@ -34,30 +26,9 @@ export default function JenkinsJob({
     subprojects: subprojectsProp,
     environment = "producao",
 }: Props) {
-    const subprojects = useMemo(
-        () => subprojectsProp ?? EMPTY_SUBPROJECTS,
-        [subprojectsProp]
-    );
-    const hasMultipleSubprojects = subprojects.length > 1;
-
-    const [selectedKey, setSelectedKey] = useState("");
+    const { subprojects, hasMultipleSubprojects, selectedKey, setSelectedKey } =
+        useSubprojectSelection(project, subprojectsProp);
     const jenkinsEnvironment = getJenkinsEnvironmentForDeploy(environment);
-
-    useEffect(() => {
-        if (!project || subprojects.length === 0) {
-            setSelectedKey("");
-            return;
-        }
-
-        if (subprojects.length === 1) {
-            setSelectedKey(subprojects[0].key);
-            return;
-        }
-
-        setSelectedKey((prev) =>
-            subprojects.some((s) => s.key === prev) ? prev : subprojects[0].key
-        );
-    }, [project, subprojects]);
 
     const query = useJenkinsMetrics({
         projectName: selectedKey,
@@ -91,32 +62,11 @@ export default function JenkinsJob({
             <div className={cn("bg-white rounded-[5px] shadow-[3px_4px_6px_0px_rgba(0,0,0,0.1)] p-5", className)}>
                 <div className="font-bold text-[14px] text-[#111827] mb-4">{title}</div>
 
-                <div className="mb-4">
-                    <div className="text-sm font-semibold text-[#111827] mb-2">Projeto</div>
-                    <Select
-                        value={selectedKey}
-                        onValueChange={(value) => setSelectedKey(value)}
-                    >
-                        <SelectTrigger
-                            size="sm"
-                            className="w-full"
-                            aria-label="Selecionar projeto"
-                        >
-                            <SelectValue placeholder="Selecione" />
-                        </SelectTrigger>
-                        <SelectContent>
-                            {subprojects.map((s) => (
-                                <SelectItem
-                                    key={s.key}
-                                    value={s.key}
-                                    className="focus:bg-[#3b82f6] focus:text-white"
-                                >
-                                    {s.label}
-                                </SelectItem>
-                            ))}
-                        </SelectContent>
-                    </Select>
-                </div>
+                <SubprojectSelect
+                    value={selectedKey}
+                    onChange={setSelectedKey}
+                    subprojects={subprojects}
+                />
 
                 <JenkinsBranchBuildsCard
                     title=""
